change balance from int to real

+++ somethingholdingmoney.cpp
@@ -1,7 +1,7 @@
 #include "somethingholdingmoney.h"
 
 SomethingHoldingMoney::SomethingHoldingMoney(int64_t id, std::string clientLogin,
-                                             int64_t initialBalance, double percents, time_t creationTime)
+                                             double initialBalance, double percents, time_t creationTime)
 {
     this->id = id;
     this->clientLogin = clientLogin;
